Feature/more parameters

+++ EndlessRunner/Assets/Scripts/UI/PlayerScoreModel.cs
@@ -6,7 +6,7 @@ public class PlayerScoreModel : MonoBehaviour
 {
     private static double score;
 
-    public int multiplier = 1;
+    public float multiplier = 1;
 
     private static int coins;
